split xmla tests
